requirements.txt
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.3
attrs==23.2.0
autoslot==2021.10.1
bcrypt==3.2.0
bottle==0.13.0
cv2-tools==2.4.0
celery==5.4.0
httpcore==0.16.3
httplib2==0.20.2
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.0
onnxruntime==1.17.3
opencv-python==4.9.0.80
pandas==2.0.3
paramiko==2.9.3
Pillow==9.3.0
plotly==5.22.0
PyMySQL==1.1.0
pyparsing==2.4.7
pyRFC3339==1.1
fast-plate-ocr==0.1.6
rich==13.8.0
tqdm==4.66.5
websocket-client==1.2.3
Werkzeug==3.0.1
yaml-1.3==0.1.0
Yaml8==0.1.2

